refactored connectors to share same abstract class
