+++ newrelic-kafka-connector/config/TelemetryEventsSinkConnector.properties
@@ -1,4 +1,4 @@
 name=TelemetryEventsSinkConnector
 topics=mytopic
 tasks.max=1
-connector.class=com.newrelic.telemetry.events.TelemetryEventsSinkConnector
+connector.class=com.newrelic.telemetry.events.EventsSinkConnector
