+++ newrelic-kafka-connector/config/TelemetryLogsSinkConnector.properties
@@ -1,5 +1,5 @@
 name=TelemetryLogsSinkConnector
-connector.class=com.newrelic.telemetry.logs.TelemetryLogsSinkConnector
+connector.class=com.newrelic.telemetry.logs.LogsSinkConnector
 topics=log_events
 tasks.max=1
 nr.timeout=10
